+++ src/euterpe/tuner/implements/tuner.py
@@ -1,4 +1,4 @@
-from ..processing.ratios import (FIVELIMIT_TUNING_RATIOS,
+from .ratios import (FIVELIMIT_TUNING_RATIOS,
                                  MEANTONE_TUNING_RATIOS,
                                  PYTHAGOREAN_TUNING_RATIOS,
                                  standard_tuning_table)
