+++ src/euterpe/utils/context.py
@@ -1,0 +1,4 @@
+from .._core.framework import Context
+
+class EuterpeContext(Context):
+    ...
