+++ src/run.py
@@ -9,7 +9,7 @@ euterpe = Euterpe("my")
 def track1():
     from pprint import pprint
 
-    ab5 = Chord("Am7(b5)/E")
+    ab5 = Chord("Abbm7(b5)/E")
     print(ab5.root, ab5.quality.name, ab5.on)
 
     pprint(Tuner.mro())
